+++ packages/authentication/modules/access/useResetPassword/index.ts
@@ -13,7 +13,8 @@ const useResetPassword = ({
   validationSchema = DEFAULT_VALIDATION_SCHEMA,
   defaultValues = DEFAULT_INITIAL_VALUES,
   ApiClass = AuthApi,
-  options,
+  enableFormApiErrors = true,
+  options = {},
 }: IUseResetPassword) => {
   const form = useForm({
     defaultValues,
@@ -26,7 +27,9 @@ const useResetPassword = ({
     ...options, // needs to be placed bellow all overridable options
     onError: (err, variables, context) => {
       options?.onError?.(err, variables, context)
-      setFormApiErrors(form, err) // this is important to show backend errors on each specific field
+      if (enableFormApiErrors) {
+        setFormApiErrors(form, err)
+      }
     },
     onSuccess: (response, variables, context) => {
       options?.onSuccess?.(response, variables, context)