+++ packages/graphql/relay.config.ts
@@ -1,0 +1,7 @@
+module.exports = {
+  src: './',
+  schema: './schema.graphql',
+  exclude: ['**/.next/**', '**/node_modules/**', '**/__generated__/**'],
+  language: 'typescript',
+  artifactDirectory: './__generated__',
+}