+++ packages/config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @baseapp-frontend/config
 
+## 2.1.1
+
+### Patch Changes
+
+- Include `relay.config.js` in the `ignorePatterns` list.
+
 ## 2.1.0
 
 ### Minor Changes